fix(#238): Adds start/suspend toolbar to context navigation
* Due to context(s) being the root(s) of the tree view, it is more
  difficult to view the start/suspend toolbar since it is only accessible
  from the Contexts View.

* Adds the toolbar to the context context so that it can be accessed when
  the specific context node is clicked on.

* Breaks out the toolbar into its own component so that it is available to
  both the Contexts View and the Context content navigation view.
